requirements.txt
Flask==1.0.2
Flask-Cors==3.0.7
numpy==1.16.1
pytest==4.2.0
pandas==0.24.1
scikit-learn==0.20.2
nltk==3.4
JPype1==0.6.3
flask_marshmallow==0.9.0
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gunicorn
Flask-PyMongo